Probability object's SetAttr method for p_value doesn't allow now to set p_value less than 0 and more than 1.

+++ common/probability_type.h
@@ -14,19 +14,16 @@ extern "C" {
 typedef struct {
     PyObject_HEAD
     PyObject            *x_attr;        /* Attributes dictionary */
-    float p_value;
+    double p_value;
 } ProbabilityObject;
 
 ProbabilityObject * newProbabilityObject(PyObject *arg);
 PyObject *ProbabilityObject_new(PyTypeObject *type, PyObject *args, PyObject *kwds);
-// int Probability_init(ProbabilityObject *self, PyObject *args, PyObject *kwds);
 void ProbabilityObject_dealloc(ProbabilityObject *self);
 int Probability_init(ProbabilityObject *self, PyObject *args, PyObject *kwds);
-//PyObject *Probability_pValSetAttr(ProbabilityObject *self, PyObject *name);
-//PyObject *Probability_pValGetAttr(ProbabilityObject *self, PyObject *name);
-
 int Probability_pValSetAttr(ProbabilityObject *self, PyObject *value);
 PyObject *Probabilit_pValGetAttr(ProbabilityObject *self,  void *Py_UNUSED(ignored));
+
 PyObject *persentage(ProbabilityObject *self, PyObject *Py_UNUSED(ignored));
 extern PyTypeObject ProbabilityType;
